Revert "Revert "Big change: topics are now referred to by name within the various hashes stored by the user object.""
And now to put the real reason in: levelViews are stored in a dictionary on the user object, there are now two steps to checking a user's permision to access a level (have they read the material, have they done the questions below). Note that all scores and view records are now stored as full hashes, not arrays (because this is more flexible).

This reverts commit ed8f5aa23678bfd3035320891ef5a5d746d24d60.